wt-push: allow specifying the date and push Linus's tags
Sometimes I need to push a build for the previous day, because
I did everything but the final push.  Let me do that by taking a
date spec as the second arg to wt-push.

Also, push Linus's tags up so that we can easily see where a build
falls relative to upstream.

Signed-off-by: Bob Copeland <me@bobcopeland.com>